Refactor MapGenerator class out of map_saver.cpp
